Move multi-select filter base to commons.filters
Relocate the shared MultiSelectListFilter base from projects.filters to
commons/filters.py and drop the module-private underscore — it is now a public,
reusable changelist-filter building block available to any app. The phase and
category filters in projects.filters subclass it unchanged.

Refs kiconiaworks/kippo
